feat: share activity component tree usage across space and entity pages
