+++ apps/web/app/space/[id]/activity/layout.tsx
