+++ apps/web/app/space/[id]/activity/loading.tsx
@@ -1,0 +1,3 @@
+import { ActivityLoading } from '~/partials/activity/activity-loading';
+
+export default ActivityLoading;
